+++ app/components/chat/Chat.client.tsx
@@ -11,6 +11,7 @@ import { workbenchStore } from '~/lib/stores/workbench';
 import { fileModificationsToHTML } from '~/utils/diff';
 import { cubicEasingFn } from '~/utils/easings';
 import { createScopedLogger, renderLogger } from '~/utils/logger';
+import { sendPromptStatusToast } from '~/utils/promptToast';
 import { BaseChat } from './BaseChat';
 
 const toastAnimation = cssTransition({
@@ -20,51 +21,6 @@ const toastAnimation = cssTransition({
 
 const logger = createScopedLogger('Chat');
 
-const sendPromptStatusToast = (message: string, rating: number, playSuccess: () => void, playFailure: () => void) => {
-  const StarRating = ({ rating }: { rating: number }) => (
-    <div className="flex items-center gap-1 mt-1">
-      {[1, 2, 3, 4, 5].map((star) => (
-        <div
-          key={star}
-          className={`text-lg ${
-            star <= rating
-              ? 'i-ph:star-fill text-yellow-400'
-              : 'i-ph:star text-gray-400'
-          }`}
-        />
-      ))}
-    </div>
-  );
-
-  const toastContent = (
-    <div className="flex flex-col">
-      <span className="font-medium">{message}</span>
-      <StarRating rating={rating} />
-    </div>
-  );
-
-  if (rating >= 3) {
-    playSuccess();
-    toast.success(toastContent, {
-      position: 'bottom-right',
-      autoClose: 3000,
-      hideProgressBar: false,
-      closeOnClick: true,
-      pauseOnHover: true,
-      draggable: true,
-    });
-  } else {
-    playFailure();
-    toast.error(toastContent, {
-      position: 'bottom-right',
-      autoClose: 3000,
-      hideProgressBar: false,
-      closeOnClick: true,
-      pauseOnHover: true,
-      draggable: true,
-    });
-  }
-};
 
 export function Chat() {
   renderLogger.trace('Chat');
